KMSCryptoUtil -> KmsCryptoClient with state
Move responsibility for regional managing of multiple kms-client to
application that uses the sqs-util-library. As AmazonSQSSecureClient
instances are created with one region in mind there is no reason for
KMSCryptoUtil to have kmsClient-per-region state. We remove an
(unneeded) optimization in order to gain the flexibility to customize
your own kmsClient (for proxy usage and such).